// src/app/modules/academicSemester/academicSemester.interface.ts
export type TMonth =
    | "January"
    | "February"
    | "March"
    | "April"
    | "May"
    | "June"
    | "July"
    | "August"
    | "September"
    | "October"
    | "November"
    | "December";

export type TAcademicSemester = {
    name: "Autum" | "Summer" | "Fall",
    code: "01" | "02" | "03",
    year: string,
    startMonth: TMonth,
    endMonth: TMonth
}